LPS-58194 SF - move user application service to product navigation
