--- modules/apps/product-menu/product-menu-my-space/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Product Menu My Space Service
-Bundle-SymbolicName: com.liferay.product.menu.my.space.service
-Bundle-Version: 1.0.0
-Include-Resource: classes
