+++ modules/apps/product-navigation/product-navigation-user/.classpath
