+++ modules/apps/product-navigation/product-navigation-user/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <projectDescription>
-	<name>product-menu-my-space-master</name>
+	<name>product-navigation-user-master</name>
 	<comment></comment>
 	<projects></projects>
 	<buildSpec>
