+++ modules/apps/product-navigation/product-navigation-user/build.gradle
