+++ modules/apps/product-navigation/product-navigation-user/snapshot.properties
