+++ modules/apps/product-navigation/product-navigation-user/src/content/Language_sk.properties
